Another go at platform agnostic tests
